requirements.txt
aiogram>=2.19
python-dotenv>=0.19.2